move Executor fixture to conftest
